+++ src/services/automower/getMowersService.ts
@@ -1,4 +1,4 @@
-import { Mower } from "../clients/automowerClient";
+import { Mower } from "../../clients/automowerClient";
 
 /**
  * A service used to retrieve the mowers associated with a Husqvarna account.
