+++ src/services/automower/pauseMowerService.ts
